+++ apps/fe/src/app/pages/release-your-worries/release-your-worries.module.scss
@@ -1,7 +1,276 @@
-/*
- * Replace this with your own classes
- *
- * e.g.
- * .container {
- * }
-*/
+$blueOxford: hsl(226, 77%, 14%);
+$blueOxford2: hsl(225, 80%, 23%);
+$blueIce: #93f1ee;
+$yellow: hsl(39, 99%, 61%);
+$indigo: hsl(246, 66%, 70%);
+$pink: #fd86d2;
+$test: hsl(224, 42%, 14%);
+
+$eggplant: hsl(290, 39%, 22%);
+$orange: hsl(28, 91%, 58%);
+$yellow2: rgba(247, 220, 171, 0.5);
+$yellow3: rgba(236, 191, 108, 0.3);
+$pink2: hsl(320, 90%, 61%);
+$teal: hsl(169, 98%, 35%);
+$teal2: hsl(169, 79%, 61%);
+
+$white: hsl(180, 57%, 97%);
+$gray1: hsl(0, 0%, 76%);
+$gray2: hsl(0, 0%, 55%);
+$gray3: hsl(0, 0%, 24%);
+$black: hsl(224, 38%, 6%);
+
+body.chat {
+  font-family: 'Tilt Neon', sans-serif;
+  color: $white;
+  overflow-x: hidden;
+  background: radial-gradient(
+      circle at 50% 50%,
+      $yellow 0%,
+      $orange 40%,
+      $yellow 70%,
+      $orange 110%
+    )
+    fixed;
+  background-blend-mode: soft-light;
+
+  .container {
+    width: 100%;
+    max-width: 800px;
+    height: 100%;
+    min-height: 100vh;
+    display: flex;
+    flex-flow: column nowrap;
+    justify-content: center;
+    align-items: center;
+    margin: 0 auto;
+    .user-top-bar {
+      position: relative;
+      top: 0;
+      left: 0;
+      height: 4rem;
+      padding: 0.5rem;
+      width: 100vw;
+      background: $blueOxford;
+      background: transparent;
+      text-align: right;
+      .full-name {
+        display: inline-block;
+        padding-right: 1rem;
+        padding-top: 0.25rem;
+        color: $white;
+      }
+      .email {
+        display: inline-block;
+        padding-right: 1rem;
+        font-size: 12px;
+        color: $gray1;
+      }
+      .signout {
+        display: inline-block;
+      }
+      img {
+        float: right;
+        width: 3rem;
+        height: auto;
+      }
+    }
+    h1 {
+      font-family: 'Permanent Marker', cursive;
+      font-size: 2rem;
+      text-align: center;
+      background: $blueOxford;
+      color: $blueIce;
+      width: 100vw;
+      padding: 1rem;
+    }
+    h2 {
+      font-family: 'Permanent Marker', cursive;
+      font-size: 2rem;
+      margin: 1.5rem 0 0 0;
+      padding: 1rem;
+      color: $blueOxford;
+      text-align: center;
+    }
+    .boodi-message {
+      width: 80%;
+      max-width: 480px;
+      font-family: 'Tilt Neon', sans-serif;
+      font-size: 16px;
+      font-style: italic;
+      margin: 24px 0 0 0;
+      padding: 0 20px;
+      line-height: 22px;
+      text-align: justify;
+      color: $blueOxford;
+    }
+    .worries-subtext {
+      width: 80%;
+      max-width: 480px;
+      font-family: 'Tilt Neon', sans-serif;
+      font-size: 16px;
+      font-style: italic;
+      margin: 10px 0 0 0;
+      padding: 0 20px;
+      line-height: 22px;
+      text-align: justify;
+      color: $blueOxford;
+    }
+    textarea {
+      width: 80%;
+      max-width: 440px;
+      height: 200px;
+      margin: 2rem 0;
+      border: 4px solid $yellow3;
+      color: $blueOxford;
+      background: $yellow2;
+      padding: 0.5rem;
+      &::placeholder {
+        color: $blueOxford;
+      }
+    }
+
+    .four-noble-truths {
+      width: 80%;
+      max-width: 440px;
+      margin-top: 2rem;
+      background: $blueOxford;
+      h2 {
+        color: $yellow;
+        margin-top: 36px;
+      }
+      .inner-content,
+      .intro-content {
+        width: 90%;
+        max-width: 400px;
+        margin: auto;
+        font-family: 'Tilt Neon', sans-serif;
+        font-size: 20px;
+        font-style: normal;
+        margin: 10px auto 0 auto;
+        padding: 20px;
+        line-height: 28px;
+        text-align: left;
+        color: $gray1;
+        h3 {
+          margin-bottom: 1rem;
+          font-weight: 800;
+        }
+        p {
+          margin-bottom: 1rem;
+        }
+      }
+    }
+    .eightfold-path {
+      width: 80%;
+      max-width: 440px;
+      background: $blueOxford;
+      margin: 2rem 0 0 0;
+      h2 {
+        color: $yellow;
+        margin-top: 36px;
+      }
+      .inner-content {
+        width: 90%;
+        max-width: 400px;
+        margin: auto;
+        font-family: 'Tilt Neon', sans-serif;
+        font-size: 20px;
+        font-style: normal;
+        margin: 10px auto 0 auto;
+        padding: 20px;
+        line-height: 28px;
+        text-align: left;
+        color: $gray1;
+        h3 {
+          margin-bottom: 1rem;
+          font-weight: 800;
+        }
+        p {
+          margin-bottom: 1rem;
+        }
+      }
+    }
+    .create-free-account {
+      height: 200px;
+      width: 80%;
+      margin: 2rem auto;
+      background: $blueIce;
+      background: transparent;
+      text-align: center;
+      p {
+        color: $blueOxford;
+        font-size: 1rem;
+        font-weight: 400;
+        line-height: 1.25rem;
+        text-align: center;
+        margin-bottom: 2rem;
+      }
+      .gimme-btn {
+        width: auto;
+      }
+    }
+  }
+  .footer-links {
+    width: 100%;
+    display: flex;
+    flex-flow: row nowrap;
+    justify-content: space-around;
+    align-items: center;
+    padding: 20px;
+    a {
+      color: $blueOxford;
+      padding: 8px 20px 10px 20px;
+      font-size: 20px;
+      height: 40px;
+      line-height: 20px;
+      &:hover {
+        background: $blueOxford;
+        color: $blueIce;
+        border: 2px solid $blueIce;
+        filter: drop-shadow(0 0 0.5rem $blueIce);
+      }
+    }
+  }
+}
+
+.chat button {
+  height: 60px;
+  border: 0;
+  padding: 1rem;
+  margin: 0;
+  font-family: 'Permanent Marker', cursive;
+  font-size: 1.5rem;
+  line-height: 1rem;
+  cursor: pointer;
+  &.primary-btn {
+    background: $blueOxford;
+    color: $blueIce;
+    border: 4px solid $blueIce;
+    &:hover {
+      filter: drop-shadow(0 0 0.5rem $blueIce);
+    }
+    &.disabled {
+      background: transparent;
+      border: 4px solid transparent;
+      color: $blueOxford;
+    }
+    &.disabled:hover {
+      filter: none;
+    }
+  }
+  &.ghost-btn {
+    font-family: 'Tilt Neon', sans-serif;
+    font-size: 1rem;
+    font-weight: 800;
+    background: transparent;
+    color: $blueOxford;
+    border: 0;
+    &:hover {
+      filter: drop-shadow(0 0 0.5rem $blueIce);
+    }
+    &.disabled {
+      background: gray;
+    }
+  }
+}
