+++ libs/auth/src/lib/components/sign-in-popup/sign-in-popup.module.scss
@@ -1,4 +1,24 @@
-@use '../../../../../../apps/fe/src/app/scss/colors';
+$blueOxford: hsl(226, 77%, 14%);
+$blueOxford2: hsl(225, 80%, 23%);
+$blueIce: #93f1ee;
+$yellow: hsl(39, 99%, 61%);
+$indigo: hsl(246, 66%, 70%);
+$pink: #fd86d2;
+$test: hsl(224, 42%, 14%);
+
+$eggplant: hsl(290, 39%, 22%);
+$orange: hsl(28, 91%, 58%);
+$yellow2: rgba(247, 220, 171, 0.5);
+$yellow3: rgba(236, 191, 108, 0.3);
+$pink2: hsl(320, 90%, 61%);
+$teal: hsl(169, 98%, 35%);
+$teal2: hsl(169, 79%, 61%);
+
+$white: hsl(180, 57%, 97%);
+$gray1: hsl(0, 0%, 76%);
+$gray2: hsl(0, 0%, 55%);
+$gray3: hsl(0, 0%, 24%);
+$black: hsl(224, 38%, 6%);
 
 .sign-in-popup {
   width: 100vw;
@@ -14,8 +34,8 @@
   .card {
     width: 80%;
     max-width: 570px;
-    background: colors.$blueOxford;
-    filter: drop-shadow(0 0 0.5rem colors.$blueIce);
+    background: $blueOxford;
+    filter: drop-shadow(0 0 0.5rem $blueIce);
     .inner-content {
       display: flex;
       flex-flow: row wrap;
@@ -40,7 +60,7 @@
         flex: 1 1 100%;
         height: 40px;
         padding: 0.5rem;
-        color: colors.$blueOxford;
+        color: $blueOxford;
       }
       input[type='email'] {
         margin-bottom: 1rem;
@@ -60,7 +80,7 @@
         font-family: 'Tilt Neon', sans-serif;
         font-size: 1rem;
         font-weight: 800;
-        color: colors.$white;
+        color: $white;
       }
       .sign-in-btn {
         flex: 1 1 100%;
@@ -68,8 +88,8 @@
         font-family: 'Tilt Neon', sans-serif;
         font-size: 1rem;
         font-weight: 800;
-        color: colors.$blueOxford;
-        background: colors.$blueIce;
+        color: $blueOxford;
+        background: $blueIce;
       }
     }
   }
@@ -86,16 +106,16 @@
   line-height: 1rem;
   cursor: pointer;
   &.primary-btn {
-    background: colors.$blueOxford;
-    color: colors.$blueIce;
-    border: 4px solid colors.$blueIce;
+    background: $blueOxford;
+    color: $blueIce;
+    border: 4px solid $blueIce;
     &:hover {
-      filter: drop-shadow(0 0 0.5rem colors.$blueIce);
+      filter: drop-shadow(0 0 0.5rem $blueIce);
     }
     &.disabled {
       background: transparent;
       border: 4px solid transparent;
-      color: colors.$blueOxford;
+      color: $blueOxford;
     }
     &.disabled:hover {
       filter: none;
@@ -106,10 +126,10 @@
     font-size: 1rem;
     font-weight: 800;
     background: transparent;
-    color: colors.$blueOxford;
+    color: $blueOxford;
     border: 0;
     &:hover {
-      filter: drop-shadow(0 0 0.5rem colors.$blueIce);
+      filter: drop-shadow(0 0 0.5rem $blueIce);
     }
     &.disabled {
       background: gray;
