--- libs/chat/README.md
@@ -1,7 +1,0 @@
-# chat
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test chat` to execute the unit tests via [Vitest](https://vitest.dev/).
