+++ libs/welcome/src/lib/welcome.module.scss
@@ -1,4 +1,24 @@
-@use '../../../../apps/fe/src/app/scss/colors';
+$blueOxford: hsl(226, 77%, 14%);
+$blueOxford2: hsl(225, 80%, 23%);
+$blueIce: #93f1ee;
+$yellow: hsl(39, 99%, 61%);
+$indigo: hsl(246, 66%, 70%);
+$pink: #fd86d2;
+$test: hsl(224, 42%, 14%);
+
+$eggplant: hsl(290, 39%, 22%);
+$orange: hsl(28, 91%, 58%);
+$yellow2: rgba(247, 220, 171, 0.5);
+$yellow3: rgba(236, 191, 108, 0.3);
+$pink2: hsl(320, 90%, 61%);
+$teal: hsl(169, 98%, 35%);
+$teal2: hsl(169, 79%, 61%);
+
+$white: hsl(180, 57%, 97%);
+$gray1: hsl(0, 0%, 76%);
+$gray2: hsl(0, 0%, 55%);
+$gray3: hsl(0, 0%, 24%);
+$black: hsl(224, 38%, 6%);
 
 body.welcome {
   background: hsl(224 71% 4%);
@@ -24,10 +44,10 @@ body.welcome {
       h1 {
         background: linear-gradient(
           150deg,
-          colors.$indigo 0%,
-          colors.$pink2 33%,
-          colors.$orange 66%,
-          colors.$yellow 100%
+          $indigo 0%,
+          $pink2 33%,
+          $orange 66%,
+          $yellow 100%
         );
         background-clip: text;
         color: transparent;
@@ -42,7 +62,7 @@ body.welcome {
       align-items: center;
       padding: 1rem;
       p {
-        color: colors.$white;
+        color: $white;
         font-family: 'Tilt Neon', sans-serif;
         font-size: 1rem;
         line-height: 1.75rem;
@@ -50,7 +70,7 @@ body.welcome {
         padding: 0.5rem 0;
       }
       b {
-        color: colors.$blueIce;
+        color: $blueIce;
       }
       img.book-lotus {
         width: 140px;
@@ -81,15 +101,15 @@ body.welcome {
   cursor: pointer;
   &.primary-btn {
     background: hsl(224 71% 4%);
-    color: colors.$blueIce;
-    border: 4px solid colors.$blueIce;
+    color: $blueIce;
+    border: 4px solid $blueIce;
     &:hover {
-      filter: drop-shadow(0 0 0.5rem colors.$blueIce);
+      filter: drop-shadow(0 0 0.5rem $blueIce);
     }
     &.disabled {
       background: transparent;
       border: 4px solid transparent;
-      color: colors.$blueOxford;
+      color: $blueOxford;
     }
     &.disabled:hover {
       filter: none;
@@ -100,10 +120,10 @@ body.welcome {
     font-size: 1rem;
     font-weight: 800;
     background: transparent;
-    color: colors.$blueOxford;
+    color: $blueOxford;
     border: 0;
     &:hover {
-      filter: drop-shadow(0 0 0.5rem colors.$blueIce);
+      filter: drop-shadow(0 0 0.5rem $blueIce);
     }
     &.disabled {
       background: gray;
